Add pytest fixtures for observation and variable data in tests
